feat: latest

+++ scripts/install-confs.sh
@@ -4,16 +4,17 @@ install_confs ()
 {
   # Install all necessary confs
   echo "I: Installing confs.."
-	mkdir ~/.config
-  mv ~/.config/starship.toml ~/.config/starship.toml${BACKUP_SFX}
-  mv ~/.bashrc ~/.bashrc${BACKUP_SFX}
-  mv ~/.bash_aliases ~/.bash_aliases${BACKUP_SFX}
-  mv ~/.tmux.conf ~/.tmux.conf${BACKUP_SFX}
-  mv ~/.config/nvim ~/.config/nvim${BACKUP_SFX}
-  mv ~/.local/share/nvim ~/.local/share/nvim${BACKUP_SFX}
-  ln -s ${DIR_GIT_REPO}/${FILE_BASHRC} ${HOME}/
-  ln -s ${DIR_GIT_REPO}/${FILE_BASHALIASES} ${HOME}/
-  ln -s ${DIR_GIT_REPO}/${FILE_TMUX} ${HOME}/
+	[[ ! -d ~/.config ]] && mkdir ~/.config || true
+  [[ ! -L ~/.config/starship.toml ]] && mv ~/.config/starship.toml ~/.config/starship.toml${BACKUP_SFX} || true
+  [[ ! -L ~/.bashrc ]] && mv ~/.bashrc ~/.bashrc${BACKUP_SFX} || true
+  [[ ! -L ~/.bash_aliases ]] && mv ~/.bash_aliases ~/.bash_aliases${BACKUP_SFX} || true
+  [[ ! -L ~/.tmux.conf ]] && mv ~/.tmux.conf ~/.tmux.conf${BACKUP_SFX} || true
+  # mv ~/.config/nvim ~/.config/nvim${BACKUP_SFX} || true
+  # mv ~/.local/share/nvim ~/.local/share/nvim${BACKUP_SFX} || true
+  ln -s ${DIR_GIT_REPO}/${FILE_BASHRC} ${HOME}/ || true
+  ln -s ${DIR_GIT_REPO}/${FILE_BASHALIASES} ${HOME}/ || true
+  ln -s ${DIR_GIT_REPO}/${FILE_TMUX} ${HOME}/ || true
+  ln -s ${DIR_GIT_REPO}/${FILE_STARSHIPTOML} ${HOME}/.config/starship.toml || true
 }
 
 install_confs
